+++ Sources/NSManagedObjectContext+Utils.swift
@@ -57,7 +57,8 @@ extension NSManagedObjectContext {
   /// **CoreDataPlus**
   ///
   /// Returns an array of objects that meet the criteria specified by a given fetch request.
-  /// @Note The Swift version returns an *Array* and for performance issues you should prefer using a NSArray* **for batched requests**: https://developer.apple.com/forums/thread/651325 .
+  /// - Note: The default Swift implementation returns an *Array* and for performance issues you should prefer using a NSArray* **for batched requests**.
+  /// - SeeAlso:https://developer.apple.com/forums/thread/651325)
   public final func fetchNSArray<T>(_ request: NSFetchRequest<T>) throws -> NSArray {
     // [...] Similarly for fetch requests with batching enabled, you do not want a Swift Array but instead an NSArray to avoid making an immediate copy of the future.
     // https://developer.apple.com/forums/thread/651325.