+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - New CoreData notifications payloads.
 - History: added new history transactions and changes fetch requests.
 - History: removed unused APIs.
+- Added a method to fetch with batched requests.
 - New batch inserts methods.
 - Removed entity and contexts observers.
 - More tests.