+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ### 3.0.0 ⭐
 
+- Added a generic *Transformer* class to implement CoreData *Transformable* attributes.
 - APIs improvements.
 - New CoreData notifications payloads.
 - History: added new history transactions and changes fetch requests.
 - History: removed unused APIs.
-- Added a method to fetch with batched requests.
+- Added a fetch method that returns *NSArray* to support batched requests.
 - New batch inserts methods.
 - Removed entity and contexts observers.
 - More tests.