+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 3.0.0 ⭐
 
-- Added a generic *Transformer* class to implement CoreData *Transformable* attributes.
+- Added a generic `NSSecureUnarchiveFromDataTransformer` subclass (*Transformer*) to easily implement CoreData *Transformable* attributes.
+- Added a generic `ValueTransfomer` closure based subclass (*DataTransformer*) to implement CoreData *Transformable* attributes.
 - APIs improvements.
 - New CoreData notifications payloads.
 - History: added new history transactions and changes fetch requests.